Separate process specific tests
